fix: speaker-stats and linting issues
